requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OCR and PDF processing
paddlepaddle-gpu==3.0.0b1
paddleocr
pymupdf==1.23.8
Pillow==10.2.0

# Translation
transformers==4.36.2
sentencepiece==0.1.99
sacremoses==0.1.1
torch==2.1.2

# PDF generation
reportlab==4.0.9
latex==0.7.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.3
